Keep using older natural_earth version
Natural Earth recently released a new version that changed some of the marine regions (e.g. #208), leading to incompatibility issues. 

Here I am using some new regionmask features to make sure that the older version is used for testing.
This is an attempt to quickly fix the CI (which has failed for the last few weeks), but I am wondering if it would be worth in the long run, to upgrade to the latest NE version.

Related question: @mathause, is there a way to identify the NE version from a regions object? 

Something similar to this:
```python
basins = regionmask.defined_regions.natural_earth_v4_1_0.ocean_basins_50
assert basins.version == 'something'
```
?
That way I could check that in my module and give a more useful errormessage

TODO:
- [ ] Mention this in the docs.