Attempt to fix test checking for user warning
Additionally to the problem with regionmask (which I started working on in #212), there seems to be an issue with a test asserting that a UserWarning is raised.

I believe the issue here is that suddenly more than one warning is raised and my old implementation did only check the first warning. 
I have now changed this to a list comparison.