Fix for CI - Keep using older natural_earth version for testing (#212)
* Keep using older natural_earth version

Natural Earth recently released a new version that changed some of the marine regions (e.g. #208), leading to incompatibility issues. 

Here I am using some new regionmask features to make sure that the older version is used for testing.
This is an attempt to quickly fix the CI (which has failed for the last few weeks), but I am wondering if it would be worth in the long run, to upgrade to the latest NE version.

Related question: @mathause, is there a way to identify the NE version from a regions object? 

Something similar to this:
```python
basins = regionmask.defined_regions.natural_earth_v4_1_0.ocean_basins_50
assert basins.version == 'something'
```
?
That way I could check that in my module and give a more useful errormessage

TODO:
- [ ] Mention this in the docs.

* Attempt to fix test checking for user warning (#213)

* Attempt to fix test checking for user warning

Additionally to the problem with regionmask (which I started working on in #212), there seems to be an issue with a test asserting that a UserWarning is raised.

I believe the issue here is that suddenly more than one warning is raised and my old implementation did only check the first warning. 
I have now changed this to a list comparison.

* Update test_drift_removal.py

* [skip-ci]